Split default implementations for pre- and post-item overlay renderers into separate classes
Move item overlay related impl classes to their own sub-package
