+++ fabric-renderer-registries-v1/src/main/java/net/fabricmc/fabric/impl/client/renderer/registry/item/DefaultCooldownOverlayProperties.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package net.fabricmc.fabric.impl.client.renderer.registry;
+package net.fabricmc.fabric.impl.client.renderer.registry.item;
 
 import net.minecraft.client.MinecraftClient;
 import net.minecraft.client.network.ClientPlayerEntity;
