+++ fabric-renderer-registries-v1/src/testmod/java/net/fabricmc/fabric/test/renderregistry/client/cooldown/FlashingCooldownProperties.java
@@ -22,7 +22,7 @@ import net.minecraft.item.ItemStack;
 import net.minecraft.util.Util;
 import net.minecraft.util.math.MathHelper;
 
-import net.fabricmc.fabric.impl.client.renderer.registry.DefaultCooldownOverlayProperties;
+import net.fabricmc.fabric.impl.client.renderer.registry.item.DefaultCooldownOverlayProperties;
 
 /**
  * Shows a full red, flashing overlay if there's more than 80% of the cooldown still remaining.
